[ALPS05444106] Clean up instructions for modifying PSTATE fields
CVE-2020-0462

Instructions for modifying the PSTATE fields which were not supported
in the older toolchains (e.g, PAN, UAO) are generated using macros.
We have so far used the normal sys_reg() helper for defining the PSTATE
fields. While this works fine, it is really difficult to correlate the
code with the Arm ARM definition.

As per Arm ARM, the PSTATE fields are defined only using Op1, Op2 fields,
with fixed values for Op0, CRn. Also the CRm field has been reserved
for the Immediate value for the instruction. So using the sys_reg()
looks quite confusing.

This patch cleans up the instruction helpers by bringing them
in line with the Arm ARM definitions to make it easier to correlate
code with the document. No functional changes.

MTK-Commit-Id: 9d2e09c78d536eafab3a3eaf284ba3eb0a889a49

Change-Id: I7c0f754ab42058f64d99bb25b55169975d738ac8
CR-Id: ALPS05444106
Feature: System Performance
Signed-off-by: mtk22228 <TzuAn.Lin@mediatek.com>