[ALPS05555734] fs: add RWF_APPEND
This is the per-I/O equivalent of O_APPEND to support atomic append
operations on any open file.

If a file is opened with O_APPEND, pwrite() ignores the offset and
always appends data to the end of the file. RWF_APPEND enables atomic
append and pwrite() with offset on a single file descriptor.

MTK-Commit-Id: e45d5058b3ea6fbe320eab4194ff9fe30d4accd6

Change-Id: I15eeb98eb183185d0fd03089ce54fdddc1950761
Signed-off-by: Jürg Billeter <j@bitron.ch>
Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>
Signed-off-by: Qilin Tan <qilin.tan@mediatek.com>
CR-Id: ALPS05555734
Feature: [Android Default] Sdcard File System