read desc
@vrdons, there is no contact information in your profile, so I will mention you here.
I noticed how you are trying to adapt my kernel changes for selene. I understand your idea, because the devices have the same hardware and the kernel sources for selene even have code for lancelot and merlin.
Earlier, I looked at the differences in their code and created this branch. You can compare it with selene-r-oss-update and see that almost all functional changes are included. But since I don't have selene, the idea of a common kernel for these devices is impossible for me. I believe you have selene. If you can pick all the necessary parts of the code for selene to work, then I can officially add your selene to the supported devices of my kernel.
Please add contact information if you wish to work on this case.